Prevent rerenders in the Triage flow
The reason you could never leave the triage flow is that it was rapidly
rerendering and setting the page content. This is because we were
creating new ClientThread instances on every render and then using that
as useEffect deps. This fixes that by putting the creating of these
threads into a useMemo. Confirmed you can leave this page now.